#78: Adjust module structure and BeheadingInterpreter docstring
