+++ tickit/adapters/interpreters/wrappers/__init__.py
@@ -1,0 +1,5 @@
+from tickit.adapters.interpreters.wrappers.beheading_interpreter import (
+    BeheadingInterpreter,
+)
+
+__all__ = ["BeheadingInterpreter"]
